% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robson1964.R
\name{robson1964}
\alias{robson1964}
\title{Robson & Whitlock (1964) truncation point estimator}
\usage{
robson1964(sd, alpha = 0.05)
}
\arguments{
\item{sd}{A \link{sighting_data} object.}

\item{alpha}{Significance level, in (0, 1), for the approximate one-sided
confidence interval.}
}
\value{
An \link{ede_estimate} object.
}
\description{
Estimates the extinction date by jackknife bias correction of the most
recent sighting. If the two most recent distinct sighting times are
\code{t[n-1]} and \code{t[n]}, the estimate is \code{t[n] + (t[n] - t[n-1])}.
}
\references{
Robson, D. S., & Whitlock, J. H. (1964). Estimation of a truncation
point. Biometrika, 51(1/2), 33-39.
}
